[TASK] Restructure the header. The two divs will create the separation between logo and navigation.
The logo needs to be wrapped in a link, as it should be clickable.

H1 instead of H2 for the logo. Hierarchical order must be correct. This way we can ensure there is always a H1 on the page.

Naming for classes should explain what it is about. Changed the classname from spin to logo.

Restructure the navigation. Without links (a href) the navigation is not clickable. Always use native elements first for interactive items.

The Hamburger Menu needs to be clickable too, but it does not cause the browser to reload. It force the browser to perform an actions. Thats why a button is the right choice here.

SVGs are code. Its better to implement them directly as code, instead of using them in an image tag.

CSS will be applied later.